# Do not reallocate whole input string when matching next token