Purely a performance refactoring.

The measured performance improvement is about -90% reduced runtime for the included test. With larger test/SQL, the speedup is even more dramatic as the original complexity was O(N^2).

Originally, the whole input string was reallocated after each token in:
- https://github.com/doctrine/sql-formatter/blob/1.4.0/src/Tokenizer.php#L790

such big temporary string was upper cased for each token matching in:
- https://github.com/doctrine/sql-formatter/blob/1.4.0/src/Tokenizer.php#L890

~and regex patterns were constructed for each token matching freshly in:~ (extracted into another PR)
- https://github.com/doctrine/sql-formatter/blob/1.4.0/src/Tokenizer.php#L874

All such operations should be avoided for linear scalability and this PR addresses that.