@mvorisek 2024-06-02
PR is finished and the performance improvement (of the "long concat" tests with new 20k elements limit) is the following:

before:
```
21.370 secs Doctrine\SqlFormatter\Tests\SqlFormatterTest::testFormatLongConcat
18.474 secs Doctrine\SqlFormatter\Tests\TokenizerTest::testTokenizeLongConcat
```

after
```
5.710 secs Doctrine\SqlFormatter\Tests\SqlFormatterTest::testFormatLongConcat
1.665 secs Doctrine\SqlFormatter\Tests\TokenizerTest::testTokenizeLongConcat
```

So the tokenization process is newly more than 10x faster and with increased query length the speedup is even bigger.