@greg0ire 2024-06-05
> Do you have a specific use case that you are solving with this lib that you need to squeeze every CPU cycle?

I'm also curious about this.