@mvorisek 2024-06-05
> > Do you have a specific use case that you are solving with this lib that you need to squeeze every CPU cycle?
> 
> I'm also curious about this.

We started using this package in https://github.com/atk4/data/pull/1216 where we use this package to format logged and debug queries. When I run all tests with this package, I noticed quite massive slowdown to I started investigating...

There were these issues:

1. constructor slowness - the class/grammar is final, we can use static property safely to build the regexes only once
2. reallocation - the solution is to use offset and `\G` instead of `^` for matching offset start
3. non-optimal regexes from lists generated - see https://3v4l.org/9jitR/rfc#vgit.master_jit to real improvent times