@greg0ire 2024-06-11
Way too many commits, way too little explanations… let's take an example among many:

> Change "^" to "\G" in regexes to match offset start 

Are we supposed to know why the hell one would want to do that?