@greg0ire 2024-06-11
> Refactor first/second char matching 

How did you refactor it and why?