@mvorisek 2024-06-11
> > Refactor first/second char matching
> 
> How did you refactor it and why?

Previously the `$str[0]` was refering to the first char, but this must be avoided, as this required many reallocations `$str = substr($str, $offset)`. Newly, we do not reallocate the string but instead of use `$offset` everywhere, this the first char is `$str[$offset]` newly, and for easier readability, I changed `$str[0]` to `$str[$offset]`. Earlier refactoring commits make sense for the next commits.