@mvorisek 2024-06-11
> Are there ways you could split this in several smaller PRs?

Yes - done now, here we focus on "no reallocation", ie. using `$offset` strictly.