@greg0ire 2024-06-11
Is 3c2a6db79e2001c2ec89b1a81a3296252cf0e97d about avoiding allocations? It seems to allocate more variables.