@mvorisek 2024-06-11
> Is [3c2a6db](https://github.com/doctrine/sql-formatter/commit/3c2a6db79e2001c2ec89b1a81a3296252cf0e97d) about avoiding allocations? It seems to allocate more variables.

Such short allocations are fast and more readable over `$str[$offset]` and `$str[$offset + 1]`. Especially, single char string is always interned thus no real allocation is done. This is thanks to internal php optimization.

However I tested it - offset is slightly faster over `substr($đtr, $offset, 1)`. I belive this is because specific opcode is used over regular function call/overhead.