@mvorisek 2024-06-11
Like really, do I have to use separate PR for using separate variable when this PR is refactoring this topic and the performance was tested?